Test of transformarEmColuna method, of class Tabuleiro. Test of transformarEmLinha method, of class Tabuleiro. E ajustes dos impactos

+++ src/br/edu/undra/MVC/modelo/impl/JogadorJogoDaVelha.java
@@ -197,8 +197,8 @@ public class JogadorJogoDaVelha<J extends Jogo> extends Jogador {
     @Override
     public boolean joga(int posicao) {
 
-        int linha = getJogo().getTabuleiro().transformarEmLinha(posicao);
-        int coluna = getJogo().getTabuleiro().transformarEmColuna(posicao);
+        int linha = (int) getJogo().getTabuleiro().transformarEmLinha(posicao);
+        int coluna = (int) getJogo().getTabuleiro().transformarEmColuna(posicao);
 
         return joga(linha, coluna);
